Add prebuilt binaries of 2xrsa, CakesROP
Neither of these compile anymore on newer toolchains, and support for
them has ceased upstream.

Both of these are unlikely to change in working behaviour, so let's
just add prebuilt binaries pulled from the v0.0.7 release and build
the remaining bits where we can.

As a result, drop the submodules in question.